refactor: Move generator class in util package
